+++ src/schemathesis/core/deserialization.py
@@ -169,8 +169,8 @@ def get_yaml_loader() -> type[yaml.SafeLoader]:
             if key_node.tag != "tag:yaml.org,2002:str":
                 key = key_node.value
             else:
-                key = self.construct_object(key_node, deep)  # type: ignore[no-untyped-call]
-            mapping[key] = self.construct_object(value_node, deep)  # type: ignore[no-untyped-call]
+                key = self.construct_object(key_node, deep)
+            mapping[key] = self.construct_object(value_node, deep)
         return mapping
 
     cls.construct_mapping = construct_mapping  # type: ignore[method-assign,assignment]